test(core): move control flow acceptance tests
Moves each control flow test (if, switch, for) into their
corresponding file.
